istekler yapildi

+++ app/src/main/java/com/hayalgucu/albawms/models/LocationModel.kt
@@ -9,7 +9,7 @@ data class LocationModel(
     @SerialName("degisti")
     val degisti: Boolean,
     @SerialName("DepoMiktari")
-    val depoMiktari: Int,
+    val depoMiktari: Double,
     @SerialName("hca_create_date")
     val hcaCreateDate: String,
     @SerialName("hca_create_user")
