+++ app/src/main/java/com/hayalgucu/albawms/models/SubLocation.kt
@@ -9,7 +9,7 @@ data class SubLocation(
     @SerialName("degisti")
     val degisti: Boolean,
     @SerialName("DepoMiktari")
-    val depoMiktari: Int,
+    val depoMiktari: Double,
     @SerialName("hca_create_date")
     val hcaCreateDate: String,
     @SerialName("hca_create_user")
